refactor: the document on this machine leaves App
The mirror of #119, split along the same line. Three things that look separate and are
one concern, because all three are "a document that lives here rather than on a server"
and all three go through the same two functions:

  .emv files    the File System Access API where it exists, a download where it does not
  IndexedDB     named projects, plus the autosave crash recovery offers on first load
  tabs          several documents open at once, each a snapshot in memory

buildData and restore arrive as functions, for the reason #119 gives. resetToEmpty joins
them for the same reason from the other end: emptying the document means clearing cuts,
tracks, the playhead, the selection, the layer cache and both media tracks, which is
App's business.

Two things the extraction forced into the open, both of them ownership questions that
were invisible while everything lived in one function:

- localNameRef is read by the server backup, for a name to fall back to, and written by
  the local save. Two hooks cannot both create it, and having the local one create it
  would mean the server hook could not be built first. App owns it and hands it to both,
  which is what fileHandleRef already needed because resetToEmpty clears it.
- Renaming a tab reached into setTabs from the JSX. It goes through renameTab now, so
  the shape of the tab list is not something the markup has to know.

Faithfulness, diffed function by function against the original: readAndRestore, doOpen,
doLocalSave, doLocalOpen, doLocalDelete and openLocalList are byte-identical, and the
whole tab block is byte-identical. doSave differs only in that one very long line is now
several - same order, same branches.

The hook baseline went 9 to 8, which is not a fix and should not read as one. The crash
recovery effect is deliberately mount-only and uses restore, so its deps must be empty;
in App that was one of the nine unexplained warnings, and here it carries a disable with
the reason written next to it. A stated reason at the site beats a number in a file, but
the count moving is bookkeeping, not repair.

App.jsx 4118 -> 3957. State 86 -> 82, refs 80 -> 76, effects 31 -> 28, top-level
functions 153 -> 141.
